+++ Packages/ai.natml.vision.blazepalm.detector/Changelog.md
@@ -1,2 +1,7 @@
+## 1.0.1
++ Added `BlazePalmDetector.Create` static method for creating the detector.
++ Removed `BlazePalmDetector` public constructor. Use `BlazePalmDetector.Create` method instead.
++ Upgraded to NatML 1.1.3.
+
 ## 1.0.0
 + First release.